requirements.txt
streamlit
pandas
requests
pydantic
pymupdf
pytesseract
Pillow
openai
pytest
pyyaml